Refactor: Unified UserLoginDefs, UserProfileDefs, and UserRegistrationDefs into UserStepDefs to follow file convention.
